Move all streams to main file
